add webpages for 3 different recipe
Chicken Adovada recipe contains the description of the recipe and the ingredients needed.
Salsa Chicken recipe contains the description of the recipe and the ingredients needed.
Torta Caprese recipe contains the description of the recipe and the ingredients needed.

All recipes contain an image that describes them.

In the main index.html file, I have added a short list to reach all recipe weppages.